done? maybe

+++ setup.sh
@@ -1,16 +1,14 @@
 #!/usr/bin/env bash
 
-# Convention:
-# Use echo to return
-# Use printf to print log
-# Wrap every variable in if statements with quotes, except for $? or something you sure that won't have white spaces
-# In fact, just wrap every variables in if statements with quotes except for $?
-# Wrap every variables that may contain white spaces or empty
-
 SCRIPT_DIR=$(dirname $(realpath $BASH_SOURCE[0]))
+
+print_help() {
+  echo "Unimplemented."
+}
 
 trim_data() {
   local data=
+
   data="${1%%#*}"                         # Remove commented
   data="${data#"${data%%[![:space:]]*}"}" # Remove leading white spaces
   data="${data%"${data##*[![:space:]]}"}" # Remove trailing white spaces
@@ -20,36 +18,47 @@ trim_data() {
 
 expand_path() {
   local path="$1"
-  [[ ${path} == "{SCRIPT_DIR}/"* ]] && path="${path//\{SCRIPT_DIR\}/${SCRIPT_DIR}}" # Replace {SCRIPT_DIR} with actual SCRIPT_DIR
-  [[ ${path} == "{HOME}/"* ]] && path="${path//\{HOME\}/${HOME}}"                   # Replace {HOME} with actual HOME
-  [[ ${path:0:1} == "~" ]] && path="${path//\~/${HOME}}"                            # Replace ~ with actual HOME
+
+  [[ ${path} = "{SCRIPT_DIR}/"* ]] && path="${path//\{SCRIPT_DIR\}/${SCRIPT_DIR}}" # Replace {SCRIPT_DIR} with actual SCRIPT_DIR
+  [[ ${path} = "{HOME}/"* ]] && path="${path//\{HOME\}/${HOME}}"                   # Replace {HOME} with actual HOME
+  [[ ${path:0:1} = "~" ]] && path="${path//\~/${HOME}}"                            # Replace ~ with actual HOME
+
   echo ${path}
 }
 
 make_symlink() {
   local source="$1"
   local destination="$2"
+  local destination_dir=$(dirname ${destination})
 
-  [[ ! -e ${source} ]] && {
+  if [[ ! -e ${source} ]]; then
     echo "Invalid source."
     return 1
-  }
+  fi
 
-  [[ "$(realpath ${destination})" = "${source}" ]] && {
+  if [[ ! -d ${destination_dir} ]]; then
+    mkdir -p ${destination_dir} || {
+      echo "Cannot create dir: ${destination_dir}"
+      return 1
+    }
+  fi
+
+  [[ "$(realpath ${destination} 2>/dev/null)" = ${source} ]] && {
     echo "Already linked, skipping."
     return 0
   }
 
-  [[ -e ${destination} ]] && {
+  if [[ -e ${destination} ]]; then
     echo "Destination existed."
-    return 2
-  }
+    return 1
+  fi
 
   output=$(ln -s ${source} ${destination} 2>&1)
-  [[ $? -ne 0 ]] && {
+
+  if [[ $? -ne 0 ]]; then
     echo "${output}"
-    return 3
-  }
+    return 1
+  fi
 
   echo "Linked ${source} -> ${destination}"
 }
@@ -57,18 +66,25 @@ make_symlink() {
 make_copy() {
   local source="$1"
   local destination="$2"
-  local source_type="$(stat --format='%F' ${source})"
+  local destination_dir=$(dirname ${destination})
+  local source_type="$(stat --format="%F" ${source} 2>/dev/null)"
 
-  [[ ! -e ${source} ]] && {
+  if [[ ! -e ${source} ]]; then
     echo "Invalid source."
     return 1
-  }
+  fi
+
+  if [[ ! -d ${destination_dir} ]]; then
+    mkdir -p ${destination_dir} || {
+      echo "Cannot create dir: ${destination_dir}"
+      return 1
+    }
+  fi
 
   if [[ -e ${destination} ]]; then
-
     local destination_type="$(stat --format='%F' ${destination})"
 
-    if [[ "${source_type}" = 'regular file' ]] && [[ "${destination_type}" = 'regular file' ]]; then
+    if [[ "${source_type}" = "regular file" && "${destination_type}" = "regular file" ]]; then
 
       local hash_source="$(md5sum ${source} 2>/dev/null)"
       local hash_destination="$(md5sum ${destination} 2>/dev/null)"
@@ -81,34 +97,30 @@ make_copy() {
     fi
 
     echo "Destination existed."
-    return 2
-
+    return 1
   fi
 
   local output=
   local status=0
 
   case "${source_type}" in
-  'regular file')
+  "regular file")
     output=$(cp ${source} ${destination} 2>&1)
     status=$?
     ;;
-
   directory)
     output=$(cp -r ${source} ${destination} 2>&1)
     status=$?
     ;;
-
   *)
     echo "Unsupported type."
-    return 3
+    return 1
     ;;
-
   esac
 
   if [[ "${status}" -ne 0 ]]; then
     echo "${output}"
-    return 4
+    return 1
   fi
 
   echo "Copied ${source} -> ${destination}"
@@ -118,15 +130,17 @@ extract_archive() {
   local archive_path=$1
   local extract_location=$2
 
-  [[ -f ${archive_path} ]] || {
+  if [[ ! -f ${archive_path} ]]; then
     echo "Archive not existed."
     return 1
-  }
+  fi
 
-  [[ -d ${extract_location} ]] || {
-    echo "Extract location is not directory or not existed."
-    return 2
-  }
+  if [[ ! -d ${extract_location} ]]; then
+    mkdir -p ${extract_location} || {
+      echo "Cannot create dir: ${extract_location}"
+      return 1
+    }
+  fi
 
   local output=
   local status=0
@@ -136,28 +150,24 @@ extract_archive() {
     output=$(unzip -qfo ${archive_path} -d ${extract_location} 2>&1)
     status=$?
     ;;
-
   *.tar.gz | *.tgz)
     output=$(tar -xzf ${archive_path} -C ${extract_location} 2>&1)
     status=$?
     ;;
-
   *.tar)
     output=$(tar -xf ${archive_path} -C ${extract_location} 2>&1)
     status=$?
     ;;
-
   *)
     echo "Unsupported archive type."
     return 2
     ;;
-
   esac
 
-  [[ "${status}" -eq 0 ]] || {
+  if [[ "${status}" -ne 0 ]]; then
     echo "${output}"
     return 3
-  }
+  fi
 
   echo "Extracted ${archive_path} -> ${extract_location}"
 }
@@ -196,24 +206,24 @@ process_manifest() {
     ;;
   *)
     echo "Unknown operation."
-    return 2
+    return 1
     ;;
   esac
 
   echo "${output}"
-  [[ "${status}" -ne 0 ]] && return 3 || return 0
+  [[ "${status}" -ne 0 ]] && return 1 || return 0
 }
 
 load_manifest() {
-  if [[ -z "$1" ]]; then
+  [[ -z "$1" ]] && {
     echo "[ERROR] Missing manifest."
     return 1
-  fi
+  }
 
-  if [[ ! -f "$1" ]]; then
+  [[ ! -f "$1" ]] && {
     echo "[ERROR] Manifest file not found."
     return 2
-  fi
+  }
 
   local line_number=0
 
@@ -225,10 +235,10 @@ load_manifest() {
 
     output=$(process_manifest "${line}")
 
-    if [[ $? -ne 0 ]]; then
+    [[ $? -ne 0 ]] && {
       printf "[ERROR] Line ${line_number}: ${output}\n"
       return 1
-    fi
+    }
 
     printf "[INFO]  Line ${line_number}: ${output}\n"
   done <"$1"
